refactor: extract task hydration from AgentWorkspace
